requirements.txt
pandas==1.0.5
python-telegram-bot==12.8
